+++ packages/core/src/bullet.ts
@@ -2,20 +2,22 @@ import config from './config'
 import builtin from './template'
 import Barrage from './barrage'
 import { math } from '@stg/utils'
-import { BulletField } from './builtin/fields'
-import { BulletJudge } from './builtin/judges'
+import { BulletJudge } from './builtin/judge'
 import { TaskHook, MountHook } from './updater'
 import Coordinate, { Point } from './coordinate'
 import CanvasPoint, { PointOptions } from './point'
 
+type BlurType = 'small' | TaskHook<Bullet, number>
 type JudgeType = 'square' | 'ortho' | 'tangent' | BulletJudge
-type FieldType = 'viewport' | 'distant' | 'timing' | BulletField
+type FieldType = 'viewport' | 'distant' | 'timing' | TaskHook<Bullet, boolean>
 
 interface BulletPoint {
   /** judge type (default: `ortho`) */
   judgeType?: JudgeType
   /** field type (default: `viewport`) */
   fieldType?: FieldType
+  /** blur function */
+  blur?: BlurType
   /** origin point */
   origin?: string | Point
   /** display layer */
@@ -40,6 +42,10 @@ export default class Bullet extends CanvasPoint<string> implements BulletPoint {
   public layer: number
   public judgeType: JudgeType
   public fieldType: FieldType
+  public blur?: TaskHook<Bullet, number>
+
+  /** @public grazing */
+  public grazing?: boolean
 
   /** @public buller id */
   public $id: number
@@ -60,6 +66,7 @@ export default class Bullet extends CanvasPoint<string> implements BulletPoint {
 
     // set bullet layer
     this.layer = options.layer === undefined ? 0 : options.layer
+    this.blur = typeof options.blur === 'string' ? builtin.blurs[options.blur] : options.blur
 
     // set judge hook and field hook
     this.judgeType = options.judgeType === undefined ? 'ortho' : options.judgeType
@@ -67,8 +74,12 @@ export default class Bullet extends CanvasPoint<string> implements BulletPoint {
     this.setTask((tick) => {
       const judge = this._resolveHook(this.judgeType, builtin.judges)
       const player = this.$barrage.$refs.player
-      if (judge && player && judge.call(this, player)) {
-        this.hitPlayer()
+      if (judge && player) {
+        if (judge.call(this, player)) {
+          this.hitPlayer()
+        } else {
+          this.grazing = judge.call(this, player, config.grazeRadius)
+        }
       }
 
       const field = this._resolveHook(this.fieldType, builtin.fields)
@@ -96,6 +107,19 @@ export default class Bullet extends CanvasPoint<string> implements BulletPoint {
 
     // set initial display
     this.display = display
+  }
+
+  render() {
+    if (!this.$context || typeof this._displayHook !== 'function') return
+    let filter = ''
+    if (typeof this.blur === 'function') {
+      const blur = this.blur(this.$tick)
+      if (blur > 0) filter += `blur(${blur}px)`
+    }
+    if (this.grazing) filter += 'sepia(0.8) contrast(1.5) hue-rotate(-0.5rad)'
+    this.$context.filter = filter || 'none'
+    this._displayHook.call(this, this.$tick)
+    this.$context.filter = 'none'
   }
 
   set display(value: string | TaskHook<this>) {