+++ packages/utils/src/index.ts
@@ -1,9 +1,11 @@
 import * as random from './random'
 import memorize from './memorize'
+import color from './color'
 import math from './math'
 
 export {
   random,
   math,
   memorize,
+  color,
 }