+++ packages/docs/guide/CHANGELOG.md
@@ -1,5 +1,27 @@
 # 更新日志
 
+## v0.3.2
+
+### 新特性
+
+- **core:** 增加了静态方法 `Coordinate.from()`, 作用是用给定的点构造坐标系.
+- **core:** CanvasPoint 增加了属性 `$barrage`, 用于访问所处的弹幕实例.
+- **core:** BulletOptions 新增了 `blur` 属性, 用于控制生成时的弹雾效果.
+- **core:** FieldOptions 增加了下列属性:
+  - `control`: 默认的自机控制模式.
+  - `borderWidth`: 自机活动范围的边框宽度.
+  - `borderHeight`: 自机活动范围的边框高度.
+- **core:** BarrageOptions, PlayerOptions 增加了 `assets` 属性, 用于声明在构造前读取的资源文件.
+- **core:** `field.getStatus` 返回对象增加了 `bulletCount` 属性, 表示当前弹幕中的子弹数量.
+- **core:** `buildFromImages` 增加一个控制自旋的参数, 当子弹旋转对称时默认时将不计算旋转以提高性能.
+- **util:** color 增加了 `scheme4`, `scheme8` 和 `scheme16` 属性作为东方 Project 风格的默认颜色方案.
+
+### 修复与其他
+
+- **core:** 修复了子弹处于斜向坐标系时出现的定位错误.
+- **core:** 修复了参考点在 mounted 钩子中无法通过 `$parent.$refs` 访问自己的问题.
+- **demo:** 在符卡 [梦幻快车] 中添加了鳞弹.
+
 ## v0.3.1
 
 ### 新特性
@@ -9,7 +31,7 @@
 - **core:** Updater 新增 `$assets`, `$config` 属性, 它们是对应模块的引用.
 - **core:** 子弹模板现在可以使用 Query 字符串, 运行时将使用类似 CSS 的优先级进行渲染.
 - **core:** BulletOptions 中的 `judgeType` 和 `fieldType` 属性都允许传入函数.
-- **core:** 新增了 `point.drawImage` 函数, 用于绘制来自资源文件的图形.
+- **core:** 新增了 `point.drawImage()` 方法, 用于绘制来自资源文件的图形.
 - **core:** 新增了 `use` 函数, 用于加载插件, 同时 bullets 模块也用插件的方式重新实现.
 - **bullets:** 增加了对东方 Project 第二代引擎 (风神录后) 的部分弹型的支持.
 - **bullets:** 新增了 `colorSeries` 对象, 包含了东方 Project 中默认的颜色系列.
@@ -57,7 +79,7 @@
 ### 新特性
 
 - **core:** 加入了默认的自机判定点颜色和大小.
-- **core:** 当调用 `scene.setBarrage` 方法时会自动清屏.
+- **core:** 当调用 `scene.setBarrage()` 方法时会自动清屏.
 - **utils:** 新增库 utils.color, 用于处理颜色.
 
 ## v0.2